+++ changelog.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file.
 - Minor update to gensnmp to compensate for invalid setting in genmon.conf
 - Added reading time to Briggs and Stratton custom controller config
 - Updated genmopeka.py add on to support more mopeka sensors and stop using old mopeka python library
+- Minor update to disable MFA if HTTPS is not enabled
 
 ## V1.18.18 -2023-07-05
 - Added new add on program gencallmebot (thanks @buschauer)